refactor(models): add improvements to models and organization
