+++ src/database/orm/mongoose/models/common/base.model.ts
@@ -1,0 +1,12 @@
+import "reflect-metadata";
+
+import * as mongoose from "mongoose";
+import { TimeStamps } from "@typegoose/typegoose/lib/defaultClasses";
+import { Prop } from "@typegoose/typegoose";
+
+import ObjectId = mongoose.Types.ObjectId;
+
+export class Base extends TimeStamps {
+  @Prop({ required: true, default: () => new ObjectId().toString() })
+  _id: string;
+}
